BUG/MEDIUM: connection: fix "bc_settings_streams_limit" typo
The keyword was correct in the doc but in the code it was spelled
with a missing 's' after 'settings', making it unavailable. Since
there was no other way to find this but reading the code, it's safe
to simply fix it and assume nobody relied on the wrong spelling.

In the worst case for older backports it can also be duplicated.

This must be backported to 3.0.

(cherry picked from commit b29560f610527acfc617e7badf990edcae799ed3)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>