BUG/MINOR: ssl: Don't allow to set NULL sni
ssl_sock_set_servername() function was documented to support NULL sni to
unset it. However, the man page of SSL_get_servername() does not mentionned
it is supported or not. And it is in fact not supported by WolfSSL and leads
to a crash if we do so.

For now, this function is never called with a NULL sni, so it better and
safer to forbid this case. Now, if the sni is NULL, the function does
nothing.

This patch could be backported to all stable versions.

(cherry picked from commit 3534efe79879daeffe52fbb5eca8955df324a981)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>
(cherry picked from commit 29451f4d2544145e32912bf0a0bbb4eb5d474dd3)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>
(cherry picked from commit d24b08625a852d9516e4273bf6083195cd51d19b)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>