MINOR: connection/ssl: Store the SNI hash value in the connection itself
When a SNI is set on a new connection, its hash is now saved in the
connection itself. To do so, a dedicated field was added into the connection
strucutre, called sni_hash. For now, this value is only used when the TLS
session is cached.

(cherry picked from commit 28654f3c9bfc1386b4e57e0f1577053dfc78f21e)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>