MINOR: ssl: Compare hashes instead of SNIs when a session is cached
This patch relies on the commit "MINOR: ssl: Store hash of the SNI for
cached TLS sessions". We now use the hash of the SNIs instead of the SNIs
themselves to know if we must update the cached SNI or not.

(cherry picked from commit 92f77cb3e6b2532a285643863d6fd26e53f4cc24)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>