MEDIUM: ssl/server: No longer store the SNI of cached TLS sessions
Thanks to the previous patch, "BUG/MEDIUM: ssl: Don't reuse TLS session
if the connection's SNI differs", it is no useless to store the SNI of
cached TLS sessions. This SNI is no longer tested and new connections
reusing a session must have the same SNI.

The main change here is for the ssl_sock_set_servername() function. It is no
longer possible to compare the SNI of the reused session with the one of the
new connection. So, the SNI is always set, with no other processing. Mainly,
the session is not destroyed when SNIs don't match. It means the commit
119a4084bf ("BUG/MEDIUM: ssl: for a handshake when server-side SNI changes")
is implicitly reverted.

It is good to note that it is unclear for me when and why the reused session
should be destroyed. Because I'm unable to reproduce any issue fixed by the
commit above.

This patch could be backported as far as 3.0 with the commit above.

(cherry picked from commit be998b590e4bb7c3f9272ca3574561f1e1294265)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>
(cherry picked from commit c936c070066c7a9f334cf6a52d7907b79f8da807)
[cf: applied on ssl_sock_ini]
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>
(cherry picked from commit c0b1e99995ed1ab0cadb8c0d451c4d0c2ec21920)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>