BUG/MINOR: mworker/cli: 'show proc' is limited by buffer size
In ticket #3204, it was reported that "show proc" is not able to display
more than 202 processes. Indeed the bufsize is 16k by default in the
master, and can't be changed anymore since 3.1.

This patch allows the 'show proc' to start again to dump when the buffer
is full, based on the timestamp of the last PID it attempted to dump.
Using pointers or count the number of processes might not be a good idea
since the list can change between calls.

Could be backported in all stable branche.
[wla: probably not worth it to backport it before 3.2]

(cherry picked from commit 594408cd612b5d80b7120d8a4f157f928ed38710)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>
(cherry picked from commit 578506048444b78da2ae4fc910c77a7b5288364c)
[wla: 3.2 still has "program" section to display]
Signed-off-by: William Lallemand <wlallemand@haproxy.com>