BUG/MEDIUM: ssl: Don't resume session for check connections
Don't attempt to use stored sessions when creating new check
connections, as the check SSL parameters might be different from the
server's ones.
This has not been proven to be a problem yet, but it doesn't mean it
can't be, and this should be backported up to 2.8 along with
dcce9369129f6ca9b8eed6b451c0e20c226af2e3 if it is.

(cherry picked from commit 420b42df1c6404956d1858e3c999e8d9288061a8)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>