BUG/MEDIUM: quic: Don't try to use hystart if not implemented
Not every CC algos implement hystart, so only call the method if it is
actually there. Failure to do so will cause crashes if hystart is on,
and the algo doesn't implement it.

This should fix github issue #3218

This should be backported up to 3.0.

(cherry picked from commit a08bc468d2c7c4bb7df84c2408d48a82f01220a0)
Signed-off-by: Christopher Faulet <cfaulet@haproxy.com>