Revert "BUG/MEDIUM: cli: State the cli have no more data to deliver if it yields"
This reverts commit 6dcc89050fbf87d08684c98fe8839cf9a69b37e7.

In 3.1 and 3.0, it breaks the interactive mode on the CLI: command
output is only delivered once the client sends a shutdown. It does
not seem that these versions are affected by the issue the patch
aimed at solving so that should be fine. If any problem would be
faced, a different fix would be needed.