Revert "BUG/MEDIUM: mux-h2: make sure not to move a dead connection to idle"
This reverts commit c18bf84b69aef178f12c6fa81f0e05748c0bf919.

As indicated in GH issue #3227, it seems that this commit is breaking
connection reuse on H2 backends. Further fixes were applied to latest
versions on top of this fix so it should probably not be backported
alone, but later fixes are more sensitive and need to be kept under
observation before being backported. Better revert this one for now,
we may take it again later once the whole series is OK.